move lat/lon stuff to float

+++ src/main/java/com/dsrts/xcopilot/model/LOCNavigationGeoPoint.java
@@ -1,7 +1,5 @@
 package com.dsrts.xcopilot.model;
 
-import com.dsrts.xcopilot.model.NavigationGeoPoint;
-
 import java.math.BigDecimal;
 
 import static java.lang.String.format;
@@ -15,8 +13,8 @@ public class LOCNavigationGeoPoint extends NavigationGeoPoint {
     public LOCNavigationGeoPoint() {}
     public LOCNavigationGeoPoint(
             Integer code,
-            BigDecimal latitude,
-            BigDecimal longitude,
+            Float latitude,
+            Float longitude,
             Integer elevationMSL,
             BigDecimal frequency,
             Integer range,
